+++ config/splat.us.dra.yaml
@@ -259,6 +259,7 @@ segments:
       - [0x628AC, c, 628AC]
       - [0x62DEC, c, 62DEC]
       - [0x62EB8, c, 62EB8]
+      - [0x66590, c, 66590]
       - [0x673C0, c, cd]
       - [0x692E8, c, 692E8]
       - [0x6BF64, c, 6BF64]
